Refactor to modularize the telemetry by components
