+++ Yuneec_SDK_iOS/YNCSDKAircraftTelemetry.h
@@ -1,5 +1,5 @@
 /*
- * YNCSDKTelemetry.h
+ * YNCSDKAircraftTelemetry.h
  * YUNEEC_SDK_IOS
  *
  * Copyright @ 2016 Yuneec.
@@ -8,49 +8,6 @@
 */
 
 #import <Foundation/Foundation.h>
-
-/**
- This class manages the real-time status of the battery component.
- */
-@interface YNCBattery : NSObject
-
-    /**
-     Battery voltage (in volts)
-     */
-    @property (nonatomic, assign) float voltageV;
-
-    /**
-     Battery remaining (as a percentage)
-     */
-    @property (nonatomic, assign) float remainingPercent;
-
-@end
-
-/**
- * This protocol provides delegate methods to subscribe to battery status updates.
- */
-@protocol YNCSDKTelemetryBatteryDelegate <NSObject>
-
-/**
- * Receives battery status updates.
- *
- * @param battery the battery object
- */
-- (void)onBatteryUpdate:(YNCBattery *)battery;
-@end
-
-/**
- This class provides a method to subscribe to battery status updates.
- */
-@interface YNCSDKTelemetryBattery : NSObject
-
- /**
- * Subscribes to battery status updates.
- *
- * @param delegate The Battery delegate object
- */
-- (void)subscribe:(id<YNCSDKTelemetryBatteryDelegate>) delegate;
-@end
 
 /**
  This class contains fields associated with the drone's position.
@@ -497,72 +454,6 @@ typedef NS_ENUM (NSInteger, YNCTelemetryFlightMode) {
 @end
 
 /**
- This class contains the drone's component health and status.
- */
-@interface YNCHealth : NSObject
-
-    /**
-     State of the gyroscope calibration
-     */
-    @property (nonatomic, assign) BOOL gyrometerCalibrationOk;
-
-    /**
-     State of the accelerometer calibration
-     */
-    @property (nonatomic, assign) BOOL accelerometerCalibrationOk;
-
-    /**
-     State of the magnetometer calibration
-     */
-    @property (nonatomic, assign) BOOL magnetometerCalibrationOk;
-
-    /**
-     State of the level calibration
-     */
-    @property (nonatomic, assign) BOOL levelCalibrationOk;
-
-    /**
-     Status of the local position
-     */
-    @property (nonatomic, assign) BOOL localPositionOk;
-
-    /**
-     Status of the global position
-     */
-    @property (nonatomic, assign) BOOL globalPositionOk;
-
-    /**
-     Status of the home position
-     */
-    @property (nonatomic, assign) BOOL homePositionOk;
-
-@end
-
-/**
- This delegate provides health status updates of the drone.
- */
-@protocol YNCSDKTelemetryHealthDelegate <NSObject>
-/**
- Receives health status updates.
- 
- @param health The health status
- */
-- (void)onHealthUpdate:(YNCHealth *)health;
-@end
-
-/**
- This class provides a method to subscribe to drone health updates.
- */
-@interface YNCSDKTelemetryHealth: NSObject
- /**
- * Subscribes to drone health updates.
- *
- * @param delegate The Drone Health delegate object
- */  
-- (void)subscribe:(id<YNCSDKTelemetryHealthDelegate>) delegate;
-@end
-
-/**
  This delegate provides armed status updates.
  */
 @protocol YNCSDKTelemetryArmedDelegate <NSObject>
@@ -586,26 +477,3 @@ typedef NS_ENUM (NSInteger, YNCTelemetryFlightMode) {
 - (void)subscribe:(id<YNCSDKTelemetryArmedDelegate>) delegate;
 @end
 
-/**
- This delegate provides the overall health status.
- */
-@protocol YNCSDKTelemetryHealthAllOkDelegate <NSObject>
-/**
- Receives overall health status updates.
- 
- @param healthAllOk true if all health flags are ok
- */
-- (void)onHealthAllOkUpdate:(BOOL)healthAllOk;
-@end
-
-/**
- This class provides a method to subscribe to overall health status updates.
- */
-@interface YNCSDKTelemetryHealthAllOk: NSObject
-/**
- * Subscribes to overall health status updates.
- *
- * @param delegate The Overall Health delegate object
- */
-- (void)subscribe:(id<YNCSDKTelemetryHealthAllOkDelegate>) delegate;
-@end
